requirements.txt
accelerate==0.16.0
chumpy==0.70
coloredlogs==15.0.1
easydict==1.10
einops==0.8.0
ema_pytorch==0.2.2
imageio==2.22.1
lxml==5.3.0
matplotlib==3.5.2
matplotlib==3.6.2
nori2==1.12.1
numpy==1.23.4
numpy==1.23.3
open3d==0.16.0
opencv_python==4.6.0.66
openmesh==1.2.1
Pillow==9.2.0
Pillow==11.0.0
plotly==5.11.0
psbody_mesh==0.4
pycocotools==2.0.4
pymeshlab==2023.12.post2
pyrender==0.1.45
pytorch3d==0.3.0
pytorch3d==0.7.4
pytorch_fid==0.3.0
PyYAML==5.4.1
PyYAML==6.0
PyYAML==6.0.2
refile==7.0.10
sapien==2.2.2
scikit_image==0.19.3
scikit_learn==1.1.2
scipy==1.14.1
setuptools==65.5.0
setuptools==65.4.1
termcolor==2.5.0
torch==1.12.1
torch_geometric==2.1.0.post1
torchgeometry==0.1.2
torchvision==0.13.1
tqdm==4.64.1
transforms3d==0.4.2
trimesh==3.15.3
